+++ jungle-examples/examples/welcome/instrumentation/bass.rs
@@ -46,9 +46,9 @@ impl Instrument for Bass {
     }
 }
 
-pub struct Thump<const NOTE: u8, const NOTE_TICK: u32, const REST_TICK: u32, const LANE_ID: u32 = 0>;
+pub struct Thump<const NOTE: u8, const NOTE_TICK: u32, const REST_TICK: u32, const LANE_ID: u8 = 0>;
 #[jungle::act]
-impl<const NOTE: u8, const NOTE_TICK: u32, const REST_TICK: u32, const LANE_ID: u32> Act
+impl<const NOTE: u8, const NOTE_TICK: u32, const REST_TICK: u32, const LANE_ID: u8> Act
     for Thump<NOTE, NOTE_TICK, REST_TICK, LANE_ID>
 {
     type Effect = Monad<Bass, LANE_ID, NOTE, NOTE_TICK, REST_TICK>;
